refactor: remove continuous/discrete subdirectories
Previously we had univariate::{continuous, discrete} and
multivariate::{continuous, discrete}. This is just a bit overkill and,
in general, I imagine most users (i.e. me) are familiar with what type
each distribution is. Furthermore, one can trivially look at the
implementors of DiscreteDistribution to find out what has been
implemented.
