--- src/multivariate/continuous/mod.rs
@@ -1,10 +1,0 @@
-pub use crate::ContinuousDistribution;
-
-mod normal;
-pub use self::normal::*;
-
-mod lognormal;
-pub use self::lognormal::*;
-
-mod dirichlet;
-pub use self::dirichlet::*;
