+++ src/multivariate/dirichlet.rs
