--- src/multivariate/discrete/mod.rs
@@ -1,4 +1,0 @@
-pub use crate::DiscreteDistribution;
-
-mod multinomial;
-pub use self::multinomial::*;
