+++ src/multivariate/mod.rs
@@ -1,4 +1,15 @@
 pub use crate::statistics::MultivariateMoments as Moments;
 
-pub mod discrete;
-pub mod continuous;
+// Continuous:
+mod normal;
+pub use self::normal::*;
+
+mod lognormal;
+pub use self::lognormal::*;
+
+mod dirichlet;
+pub use self::dirichlet::*;
+
+// Discrete:
+mod multinomial;
+pub use self::multinomial::*;
