+++ src/multivariate/multinomial.rs
