+++ src/multivariate/normal/bivariate.rs
@@ -87,6 +87,10 @@ impl Distribution for BivariateNormal {
         TwoSpace::new([Reals, Reals])
     }
 
+    fn cdf(&self, x: [f64; 2]) -> Probability {
+        unimplemented!()
+    }
+
     fn sample<R: Rng + ?Sized>(&self, rng: &mut R) -> [f64; 2] {
         let z1: f64 = rng.sample(RandSN);
         let z2: f64 = rng.sample(RandSN);
