+++ src/univariate/arcsine.rs
