+++ src/univariate/bernoulli.rs
