+++ src/univariate/beta.rs
