+++ src/univariate/beta_binomial.rs
