+++ src/univariate/categorical.rs
