+++ src/univariate/cauchy.rs
