+++ src/univariate/cosine.rs
