+++ src/univariate/exponential.rs
