+++ src/univariate/folded_normal.rs
