+++ src/univariate/frechet.rs
