+++ src/univariate/gamma.rs
