+++ src/univariate/geometric.rs
