+++ src/univariate/gumbel.rs
