+++ src/univariate/inverse_gamma.rs
