+++ src/univariate/inverse_normal.rs
