+++ src/univariate/laplace.rs
