+++ src/univariate/normal.rs
