+++ src/univariate/pareto.rs
