+++ src/univariate/poisson.rs
