+++ src/univariate/rayleigh.rs
