+++ src/univariate/triangular.rs
